[#3558] AlignmentCorrector extraction
