+++ lib/rubocop/cop/layout/align_array.rb
@@ -21,13 +21,17 @@ module RuboCop
       #        'forrest',
       #        'run']
       class AlignArray < Cop
-        include AutocorrectAlignment
+        include Alignment
 
         MSG = 'Align the elements of an array literal if they span more ' \
               'than one line.'.freeze
 
         def on_array(node)
           check_alignment(node.children)
+        end
+
+        def autocorrect(node)
+          AlignmentCorrector.correct(processed_source, node, column_delta)
         end
       end
     end
